added directory moved files and changed paths in makefile
